better seperation of code. now the api client lives in metta.common and there is a heatmap_widget.util module
