+++ common/src/metta/common/client/__init__.py
@@ -1,0 +1,5 @@
+# Metta AI api client
+
+from .metta_client import MettaAPIClient
+
+__all__ = ["MettaAPIClient"]
