+++ experiments/notebooks/utils/heatmap_widget/__init__.py
@@ -6,13 +6,10 @@ from .heatmap_widget.HeatmapWidget import (
     create_heatmap_widget,
     create_multi_metric_demo,
 )
-from .heatmap_widget.metta_client import MettaAPIClient, fetch_real_heatmap_data
 
 __all__ = [
     "HeatmapWidget",
     "create_demo_heatmap",
     "create_heatmap_widget",
     "create_multi_metric_demo",
-    "MettaAPIClient",
-    "fetch_real_heatmap_data",
 ]
